requirements.txt
activate==0.2.1
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
cffi==1.14.3
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gunicorn==20.0.4
importlib-metadata==2.0.0
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.0
opencv-python==4.4.0.46
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.5
py==1.9.0
pycparser==2.20
pydicom==2.1.2
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pyzmq==19.0.1
qtconsole==4.7.7
QtPy==1.9.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0
efficientnet-pytorch==0.7.0